+++ subsys/bluetooth/audio/shell/bap_scan_delegator.c
@@ -31,7 +31,7 @@
 #include <zephyr/types.h>
 
 #include <audio/bap_internal.h>
-#include "shell/bt.h"
+#include "host/shell/bt.h"
 
 #define PA_SYNC_INTERVAL_TO_TIMEOUT_RATIO 20 /* Set the timeout relative to interval */
 #define PA_SYNC_SKIP              5
