+++ subsys/bluetooth/audio/shell/csip_set_coordinator.c
@@ -31,7 +31,7 @@
 #include <zephyr/types.h>
 #include <zephyr/shell/shell.h>
 
-#include "shell/bt.h"
+#include "host/shell/bt.h"
 
 static uint8_t members_found;
 static struct k_work_delayable discover_members_timer;
