+++ subsys/bluetooth/audio/shell/media_controller.c
@@ -28,7 +28,7 @@
 
 #include "../media_proxy_internal.h" /* For MPL_NO_TRACK_ID - TODO: Fix */
 
-#include "shell/bt.h"
+#include "host/shell/bt.h"
 
 LOG_MODULE_REGISTER(bt_media_controller_shell, CONFIG_BT_MCS_LOG_LEVEL);
