+++ subsys/bluetooth/audio/shell/pbp.c
@@ -23,7 +23,7 @@
 #include <zephyr/sys/__assert.h>
 #include <zephyr/sys/util.h>
 
-#include "shell/bt.h"
+#include "host/shell/bt.h"
 
 #define PBS_DEMO                'P', 'B', 'P'
