+++ subsys/bluetooth/host/CMakeLists.txt
@@ -4,6 +4,7 @@ zephyr_library()
 zephyr_library_link_libraries(subsys__bluetooth)
 
 add_subdirectory_ifdef(CONFIG_BT_CLASSIC classic)
+add_subdirectory_ifdef(CONFIG_BT_SHELL shell)
 
 zephyr_library_sources_ifdef(CONFIG_BT_HCI_RAW          hci_raw.c hci_common.c)
 zephyr_library_sources_ifdef(CONFIG_BT_MONITOR          monitor.c)
