+++ subsys/bluetooth/host/classic/CMakeLists.txt
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
+add_subdirectory_ifdef(CONFIG_BT_SHELL shell)
+
 zephyr_library()
 zephyr_library_link_libraries(subsys__bluetooth)
