+++ subsys/bluetooth/host/classic/shell/CMakeLists.txt
@@ -1,0 +1,6 @@
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources(bredr.c)
+zephyr_library_sources_ifdef(CONFIG_BT_RFCOMM rfcomm.c)
+zephyr_library_sources_ifdef(CONFIG_BT_A2DP a2dp.c)
