+++ subsys/bluetooth/mesh/shell/hci.c
