+++ subsys/bluetooth/services/ias/CMakeLists.txt
@@ -1,5 +1,7 @@
 # SPDX-License-Identifier: Apache-2.0
 
+add_subdirectory_ifdef(CONFIG_BT_SHELL shell)
+
 zephyr_sources_ifdef(
         CONFIG_BT_IAS
         ias.c)
