+++ subsys/bluetooth/services/ias/shell/CMakeLists.txt
@@ -1,0 +1,5 @@
+# SPDX-License-Identifier: Apache-2.0
+
+zephyr_library()
+zephyr_library_sources_ifdef(CONFIG_BT_IAS ias.c)
+zephyr_library_sources_ifdef(CONFIG_BT_IAS_CLIENT ias_client.c)
