--- subsys/bluetooth/shell/CMakeLists.txt
@@ -1,53 +1,0 @@
-# SPDX-License-Identifier: Apache-2.0
-
-zephyr_library()
-zephyr_library_sources(
-  bt.c
-  hci.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_CONN
-  gatt.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_CLASSIC
-  bredr.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_L2CAP_DYNAMIC_CHANNEL
-  l2cap.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_RFCOMM
-  rfcomm.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_A2DP
-  a2dp.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_ISO
-  iso.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_CHANNEL_SOUNDING
-  cs.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_IAS
-  ias.c
-  )
-zephyr_library_sources_ifdef(
-  CONFIG_BT_IAS_CLIENT
-  ias_client.c
-  )
-
-if(CONFIG_BT_CTLR AND CONFIG_BT_LL_SW_SPLIT)
-  zephyr_library_sources(
-    ll.c
-    ticker.c
-    )
-  zephyr_include_directories(
-    ${ZEPHYR_BASE}/subsys/bluetooth/controller/ll_sw/nordic
-    )
-endif()
